requirements.txt
appdirs==1.4.4
argcomplete==3.5.3
art==6.4
importlib_resources==6.5.2
numpy==2.2.2
pillow==11.1.0
pyscard==2.2.1
python-barcode==0.15.1
python-escpos==3.1
pyusb==1.3.1
PyYAML==6.0.2
qrcode==8.0
setuptools==75.8.0
six==1.17.0
wheel==0.45.1
